refactor: auth 패키지명 persentaton → presentation 수정
